Move Ray Serve integration coverage under serve tests
