+++ tests/core/test_serve.py
@@ -20,10 +20,7 @@ from pytest_httpserver import HTTPServer
 
 LLMConfig = pytest.importorskip("ray.serve.llm", reason="ray[serve] not installed").LLMConfig
 
-from nemo_curator.core.serve import InferenceModelConfig, InferenceServer, is_inference_server_active  # noqa: E402
-
-INTEGRATION_TEST_MODEL = "HuggingFaceTB/SmolLM2-135M-Instruct"  # pragma: allowlist secret
-INTEGRATION_TEST_MODEL_2 = "HuggingFaceTB/SmolLM-135M-Instruct"  # pragma: allowlist secret
+from nemo_curator.core.serve import InferenceModelConfig, InferenceServer  # noqa: E402
 
 
 class TestInferenceModelConfig:
@@ -178,119 +175,3 @@ class TestInferenceServer:
         fresh.stop()
         fresh.stop()
         assert fresh._started is False
-
-
-# ---------------------------------------------------------------------------
-# Integration tests — real Ray Serve + vLLM
-# ---------------------------------------------------------------------------
-@pytest.fixture(scope="class")
-def model_server(shared_ray_cluster: str) -> InferenceServer:  # noqa: ARG001
-    """Start InferenceServer once for all integration tests.
-
-    Uses enforce_eager=True to skip torch.compile and CUDA graph capture,
-    cutting vLLM startup from ~30s to ~5s.
-    """
-    config = InferenceModelConfig(
-        model_identifier=INTEGRATION_TEST_MODEL,
-        deployment_config={
-            "autoscaling_config": {"min_replicas": 1, "max_replicas": 1},
-        },
-        engine_kwargs={
-            "tensor_parallel_size": 1,
-            "max_model_len": 512,
-            "enforce_eager": True,
-        },
-    )
-
-    server = InferenceServer(models=[config], health_check_timeout_s=600)
-    server.start()
-
-    yield server
-
-    server.stop()
-
-
-@pytest.mark.gpu
-@pytest.mark.usefixtures("model_server")
-class TestInferenceServerIntegration:
-    """Full lifecycle tests against a real InferenceServer started once for the class."""
-
-    def test_is_active_and_queryable(self, model_server: InferenceServer) -> None:
-        """Server is active, lists models, and responds to chat completions."""
-        from openai import OpenAI
-
-        assert is_inference_server_active()
-        assert model_server._started is True
-
-        client = OpenAI(base_url=model_server.endpoint, api_key="na")
-
-        # /v1/models lists our model
-        model_ids = [m.id for m in client.models.list()]
-        assert INTEGRATION_TEST_MODEL in model_ids
-
-        # Chat completion returns a non-empty response
-        response = client.chat.completions.create(
-            model=INTEGRATION_TEST_MODEL,
-            messages=[{"role": "user", "content": "Say hello in one word."}],
-            max_tokens=16,
-            temperature=0.0,
-        )
-        assert len(response.choices) > 0
-        assert len(response.choices[0].message.content) > 0
-
-    def test_second_start_rejected(self, model_server: InferenceServer) -> None:
-        """Cannot start a second InferenceServer while one is already active."""
-        server2 = InferenceServer(
-            models=[
-                InferenceModelConfig(
-                    model_identifier=INTEGRATION_TEST_MODEL_2,
-                    deployment_config={"autoscaling_config": {"min_replicas": 1, "max_replicas": 1}},
-                    engine_kwargs={"tensor_parallel_size": 1, "max_model_len": 512, "enforce_eager": True},
-                )
-            ],
-            health_check_timeout_s=600,
-        )
-        with pytest.raises(RuntimeError, match="already active"):
-            server2.start()
-
-        # First server is still healthy and unaffected
-        from openai import OpenAI
-
-        client = OpenAI(base_url=model_server.endpoint, api_key="na")
-        assert INTEGRATION_TEST_MODEL in {m.id for m in client.models.list()}
-
-    def test_restart_after_stop(self, model_server: InferenceServer) -> None:
-        """A new InferenceServer starts cleanly after the previous one is stopped.
-
-        stop() calls serve.shutdown(), so start() must recreate the
-        controller and HTTP proxy from scratch.  This test must run last
-        in the class — it stops the shared fixture's server and starts a
-        replacement.
-        """
-        from openai import OpenAI
-
-        # Stop the fixture's server
-        model_server.stop()
-        assert not is_inference_server_active()
-
-        # Start a fresh server from scratch (new controller + proxy)
-        config = InferenceModelConfig(
-            model_identifier=INTEGRATION_TEST_MODEL,
-            deployment_config={"autoscaling_config": {"min_replicas": 1, "max_replicas": 1}},
-            engine_kwargs={"tensor_parallel_size": 1, "max_model_len": 512, "enforce_eager": True},
-        )
-        server2 = InferenceServer(models=[config], health_check_timeout_s=600)
-        server2.start()
-
-        client = OpenAI(base_url=server2.endpoint, api_key="na")
-        assert INTEGRATION_TEST_MODEL in {m.id for m in client.models.list()}
-
-        response = client.chat.completions.create(
-            model=INTEGRATION_TEST_MODEL,
-            messages=[{"role": "user", "content": "Say hello in one word."}],
-            max_tokens=16,
-            temperature=0.0,
-        )
-        assert len(response.choices[0].message.content) > 0
-
-        server2.stop()
